+++ lib/src/presentation/widgets/attachments/pdf_attachment_loader.dart
@@ -1,9 +1,11 @@
+import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 
 import '../../../domain/models/message_attachment.dart';
 import '../../theme/chat_theme.dart';
+import '../../utils/attachment_utils.dart';
 import 'pdf_document_card.dart';
 
 class PdfAttachmentLoader extends StatelessWidget {
@@ -16,12 +18,19 @@ class PdfAttachmentLoader extends StatelessWidget {
   final MessageAttachment attachment;
   final Future<List<int>> Function(String storagePath) resolveBytes;
 
+  Future<List<int>> _loadBytes() async {
+    if (isLocalAttachmentPath(attachment.storagePath)) {
+      return File(attachment.storagePath).readAsBytes();
+    }
+    return resolveBytes(attachment.storagePath);
+  }
+
   @override
   Widget build(BuildContext context) {
     final theme = chatThemeOf(context);
 
     return FutureBuilder<List<int>>(
-      future: resolveBytes(attachment.storagePath),
+      future: _loadBytes(),
       builder: (context, snapshot) {
         if (snapshot.connectionState != ConnectionState.done) {
           return SizedBox(